pyproject.toml
[project]
name = "talon-gaze-ocr"
version = "0.1.0"
description = "Talon scripts to enable advanced cursor control using eye tracking and text recognition (OCR)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
dev = [
    "click>=8.1.0",
    "line-profiler>=5.0.0",
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pytest>=8.0.0",
    "pytest-benchmark>=4.0.0",
]

[tool.ruff]
extend-exclude = [".subtrees/rapidfuzz"]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["E501", "SIM105", "SIM116", "UP045"]

[tool.pytest.ini_options]
testpaths = [".tests"]

[tool.pyright]
pythonVersion = "3.11"
# Talon action class methods don't use self so ignore the associated errors and warnings
reportGeneralTypeIssues = false
reportSelfClsParameterName = false
# Talon can't be installed in CI so ignore source errors
reportMissingImports = false
extraPaths = [".subtrees/gaze-ocr", ".subtrees/screen-ocr", ".subtrees/rapidfuzz/src"]
# Override auto-exclude for owned subtrees, but keep rapidfuzz excluded
exclude = [".subtrees/rapidfuzz"]
